Server - FastApi + Supabase/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.18.1
pydantic>=2.11.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
Pillow==10.0.1